# openapi-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/payments/webhook": {
      "post": {
        "operationId": "PaymentsController_webhook",
        "parameters": [
          {
            "name": "stripe-signature",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Payments"
        ]
      }
    },
...
    "/users/{id}/orders": {
      "get": {
        "operationId": "UsersController_getUserOrders",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Users"
        ]
      }
    },
...
  },
  "info": {
    "title": "Demomart API",
    "description": "API documentation for Demomart",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
...
}